requirements.txt
Django~=4.0.3
django-filter~=21.1
djangorestframework~=3.13.1
drf_yasg~=1.20.0
freezegun~=1.2.0
gunicorn~=20.1
model-bakery~=1.4.0
psycopg2~=2.9.3
pytest~=7.0.1
pytest-django~=4.5.2
python-decouple~=3.6